+++ corehq/apps/hqadmin/views.py
@@ -54,7 +54,7 @@ from corehq.apps.hqadmin.reporting.exceptions import HistoTypeNotFoundException
 from corehq.apps.hqwebapp.views import BaseSectionPageView
 from corehq.apps.locations.models import SQLLocation
 from corehq.apps.ota.views import get_restore_response, get_restore_params
-from corehq.apps.style.decorators import use_datatables, use_jquery_ui, \
+from corehq.apps.hqwebapp.decorators import use_datatables, use_jquery_ui, \
     use_nvd3_v3
 from corehq.apps.users.models import CommCareUser, WebUser, CouchUser
 from corehq.apps.users.util import format_username
