+++ corehq/apps/reports/standard/project_health.py
@@ -4,7 +4,7 @@ from django.utils.translation import ugettext as _, ugettext_lazy
 from corehq.apps.data_analytics.models import MALTRow
 from corehq.apps.domain.models import Domain
 from corehq.apps.reports.standard import ProjectReport
-from corehq.apps.style.decorators import use_nvd3
+from corehq.apps.hqwebapp.decorators import use_nvd3
 from corehq.apps.users.util import raw_username
 from dimagi.ext import jsonobject
 from dimagi.utils.dates import add_months
