+++ corehq/apps/styleguide/templates/styleguide/docs/controls_demo/views.html
@@ -9,7 +9,7 @@
 <span class="kn">from</span> <span class="nn">django.utils.decorators</span> <span class="kn">import</span> <span class="n">method_decorator</span>
 <span class="kn">from</span> <span class="nn">django.utils.translation</span> <span class="kn">import</span> <span class="n">ugettext_noop</span><span class="p">,</span> <span class="n">ugettext</span> <span class="k">as</span> <span class="n">_</span>
 <span class="kn">from</span> <span class="nn">corehq.apps.domain.decorators</span> <span class="kn">import</span> <span class="n">login_required</span>
-<span class="kn">from</span> <span class="nn">corehq.apps.style.decorators</span> <span class="kn">import</span> <span class="n">use_select2</span>
+<span class="kn">from</span> <span class="nn">corehq.apps.hqwebapp.decorators</span> <span class="kn">import</span> <span class="n">use_select2</span>
 <span class="kn">from</span> <span class="nn">corehq.apps.style.views</span> <span class="kn">import</span> <span class="n">BaseSectionPageView</span>
 <span class="kn">from</span> <span class="nn">corehq.apps.styleguide.examples.controls_demo.forms</span> <span class="kn">import</span> <span class="p">(</span>
     <span class="n">SelectControlDemoForm</span><span class="p">,</span>
