+++ corehq/apps/userreports/tests/data/report_builder_v2_diffs/views.py.diff.txt
@@ -66,7 +66,7 @@
 +from corehq.apps.locations.permissions import conditionally_location_safe
  from corehq.apps.domain.views import BaseDomainView
  from corehq.apps.reports.dispatcher import cls_to_view_login_and_domain
- from corehq.apps.style.decorators import (
+ from corehq.apps.hqwebapp.decorators import (
 @@ -31,12 +53,19 @@
      use_daterangepicker,
      use_datatables,
