+++ corehq/apps/userreports/views.py
@@ -48,7 +48,7 @@ from corehq.apps.domain.decorators import login_and_domain_required, login_or_ba
 from corehq.apps.locations.permissions import conditionally_location_safe
 from corehq.apps.domain.views import BaseDomainView
 from corehq.apps.reports.dispatcher import cls_to_view_login_and_domain
-from corehq.apps.style.decorators import (
+from corehq.apps.hqwebapp.decorators import (
     use_select2,
     use_daterangepicker,
     use_datatables,
