+++ custom/ewsghana/reports/maps.py
@@ -7,7 +7,7 @@ from corehq.apps.reports.commtrack.const import STOCK_SECTION_TYPE
 from corehq.apps.reports.commtrack.data_sources import StockStatusBySupplyPointDataSource
 from corehq.apps.reports.commtrack.maps import StockStatusMapReport
 from corehq.apps.reports.standard import CustomProjectReport
-from corehq.apps.style.decorators import maps_prefer_canvas, use_maps
+from corehq.apps.hqwebapp.decorators import maps_prefer_canvas, use_maps
 from custom.ewsghana.utils import get_country_id, filter_slugs_by_role
 from dimagi.utils.decorators.memoized import memoized
