+++ custom/ilsgateway/tanzania/reports/alerts.py
@@ -6,7 +6,7 @@ from corehq.apps.reports.datatables import DataTablesHeader, DataTablesColumn
 from corehq.apps.reports.filters.fixtures import AsyncLocationFilter
 from corehq.apps.reports.generic import GenericTabularReport
 from corehq.apps.reports.standard import CustomProjectReport, ProjectReportParametersMixin
-from corehq.apps.style.decorators import use_nvd3
+from corehq.apps.hqwebapp.decorators import use_nvd3
 from couchexport.models import Format
 from custom.ilsgateway.filters import ILSDateFilter
 from custom.ilsgateway.models import Alert
