+++ custom/ilsgateway/views.py
@@ -15,7 +15,7 @@ from corehq.apps.products.models import SQLProduct
 from corehq.apps.domain.views import BaseDomainView, DomainViewMixin
 from corehq.apps.locations.models import SQLLocation
 from corehq.apps.sms.models import SMS, INCOMING, OUTGOING
-from corehq.apps.style.decorators import use_datatables
+from corehq.apps.hqwebapp.decorators import use_datatables
 from corehq.apps.users.models import CommCareUser, WebUser, UserRole
 from django.http import HttpResponse
 from django.utils.translation import ugettext_noop
